moved branch_stats() fn to TreeSequence; others to test file
